- Start to refactor directory structure to have a better organization - Added more MaterialUi factories for edit and details type
