+++ demo/components/LiveSchemaEditor.js
@@ -4,8 +4,8 @@ import presets from '../presets';
 import _ from 'underscore';
 import psjon from '../../package.json';
 import AutoForm from '../../src/AutoForm';
-import editComponentFactory from '../../src/DefaultEditComponentFactory';
-import detailsComponentFactory from '../../src/DefaultDetailsComponentFactory';
+import editComponentFactory from '../../src/factory/bootstrap/BootstrapEditComponentFactory';
+import detailsComponentFactory from '../../src/factory/bootstrap/BootstrapDetailsComponentFactory';
 import {Alert, Badge} from 'react-bootstrap';
 import moment from 'moment';
 import numbro from 'numbro';
