+++ demo/less/styles.less
@@ -3,7 +3,7 @@
 @import '../../node_modules/react-widgets/dist/css/react-widgets.css';
 @import '../../node_modules/react-select/dist/react-select.css';
 
-@import '../../src/less/styles';
-@import '../../src/less/styles-defaultfactories';
+@import '../../src/less/bootstrap/styles';
+@import '../../src/less/bootstrap/styles-defaultfactories';
 
 @import './LiveSchemaEditor.less';
