--- src/factory/BootstrapComponentFactory.js
@@ -1,1 +1,0 @@
-
