+++ src/factory/ComponentFactory.js
