+++ src/factory/material-ui/helper/MaterialUiComponentFactoryHelper.js
@@ -1,0 +1,28 @@
+// import TextBox from './../../../components/material-ui/fieldComponents/TextBox';
+// import Select from './../../../components/material-ui/fieldComponents/Select';
+// import TextArea from './../../../components/material-ui/fieldComponents/TextArea';
+// import Group from './../../../components/material-ui/groupComponents/Group';
+// import ArrayContainer from './../../../components/material-ui/fieldComponents/ArrayContainer';
+// import DateTimePicker from './../../../components/material-ui/fieldComponents/DateTimePicker';
+// import Lookup from './../../../components/material-ui/fieldComponents/Lookup';
+// import Static from './../../../components/material-ui/fieldComponents/Static';
+// import FieldGroup from './../../../components/material-ui/fieldComponents/FieldGroup';
+// import Checkbox from './../../../components/material-ui/fieldComponents/CheckBox';
+// import Radio from './../../../components/material-ui/fieldComponents/Radio';
+
+export function registerBaseComponents(componentFactory) {
+    if (!componentFactory) throw Error('\'componentFactory\' should be truthy');
+
+    // componentFactory.registerFieldComponent('TextBox', ['string', 'int', 'float', 'datetime', 'date', 'time'], TextBox);
+    // componentFactory.registerFieldComponent('Select', ['string'], Select);
+    // componentFactory.registerFieldComponent('Radio', ['string'], Radio);
+    // componentFactory.registerFieldComponent('Lookup', ['string'], Lookup);
+    // componentFactory.registerFieldComponent('TextArea', ['string'], TextArea);
+    // componentFactory.registerFieldComponent('ArrayContainer', ['array'], ArrayContainer);
+    // componentFactory.registerFieldComponent('DateTimePicker', ['datetime', 'date', 'time'], DateTimePicker);
+    // componentFactory.registerFieldComponent('Checkbox', ['bool'], Checkbox);
+    // componentFactory.registerFieldComponent('Static', ['string', 'int', 'float', 'datetime', 'date', 'time', 'bool'], Static);
+    // componentFactory.registerFieldComponent('FieldGroup', ['group'], FieldGroup);
+    //
+    // componentFactory.registerGroupComponent('Group', Group);
+}
