+++ src/index.js
@@ -2,9 +2,9 @@
 export AutoForm from './AutoForm.js';
 
 // component factories
-export ComponentFactory from './ComponentFactory.js';
-export DefaultEditComponentFactory from './DefaultEditComponentFactory';
-export DefaultDetailsComponentFactory from './DefaultDetailsComponentFactory';
+export ComponentFactory from './factory/ComponentFactory.js';
+export DefaultEditComponentFactory from './factory/bootstrap/BootstrapEditComponentFactory';
+export DefaultDetailsComponentFactory from './factory/bootstrap/BootstrapDetailsComponentFactory';
 
 //localizers
 export momentLocalizer from './lib/localization/momentLocalizer';
