+++ src/less/bootstrap/styles-defaultfactories.less
@@ -1,4 +1,4 @@
 // third party
 @import './defaultFactories/Lookup';
-@import 'defaultFactories/DateTimePicker';
+@import './defaultFactories/DateTimePicker';
 @import './defaultFactories/Checkbox';
